requirements.txt
Django==4.1.7
django-extensions==3.2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.0
freezegun==1.2.2
schema==0.7.5